+++ src/main/java/com/trivago/rta/constants/Status.java
@@ -20,7 +20,7 @@ package com.trivago.rta.constants;
  * Enum to manage all states for steps and scenarios.
  */
 public enum Status {
-    PASSED("passed"), FAILED("failed"), SKIPPED("skipped"), UNDEFINED("undefined");
+    PASSED("passed"), FAILED("failed"), SKIPPED("skipped"), PENDING("pending"), UNDEFINED("undefined");
 
     private final String status;