+++ src/main/java/com/trivago/rta/json/pojo/Step.java
@@ -59,22 +59,6 @@ public class Step extends ResultMatch {
         this.embeddings = embeddings;
     }
 
-    public Status getStatus() {
-        return Status.fromString(getResult().getStatus());
-    }
-
-    public boolean isFailed(){
-        return getStatus() == Status.FAILED;
-    }
-
-    public boolean isPassed(){
-        return getStatus() == Status.PASSED;
-    }
-
-    public boolean isSkipped(){
-        return getStatus() == Status.SKIPPED || getStatus() == Status.UNDEFINED;
-    }
-
     @Override
     public String toString() {
         return "Step{" +