+++ src/main/java/com/trivago/rta/rendering/pages/DetailPageRenderer.java
@@ -45,6 +45,8 @@ public class DetailPageRenderer extends PageRenderer {
         addCurrentDateToReportDetails(reportDetails);
         detailPageCollection.setReportDetails(reportDetails);
 
+        System.out.println("Rendering " + detailPageCollection.getElement().getName());
+
         Writer stringWriter = new StringWriter();
         try {
             template.process(detailPageCollection, stringWriter);