+++ CHANGELOG.md
@@ -12,6 +12,15 @@ Back to [Readme](README.md).
 - Support for non-image attachments
 - Increased unit test coverage
 
+## [0.0.5] - 2017-11-20
+
+### Fixed
+
+### Added
+
+- Before and After hooks are displayed in the report
+- Total test time is shown in the start page
+
 ## [0.0.4] - 2017-11-20
 
 ### Added
@@ -34,7 +43,8 @@ Back to [Readme](README.md).
 
 Initial project version on GitHub and Maven Central.
 
-[Unreleased]: https://github.com/trivago/cluecumber-report-plugin/compare/0.0.4...HEAD
+[Unreleased]: https://github.com/trivago/cluecumber-report-plugin/compare/0.0.5...HEAD
+[0.0.5]: https://github.com/trivago/cluecumber-report-plugin/tree/0.0.5
 [0.0.4]: https://github.com/trivago/cluecumber-report-plugin/tree/0.0.4
 [0.0.3]: https://github.com/trivago/cluecumber-report-plugin/tree/0.0.3
 [0.0.2]: https://github.com/trivago/cluecumber-report-plugin/tree/0.0.2